Use CGFloat instead of float for 64bit compatibility

+++ RockPack/GKImageCropViewController.m
@@ -196,7 +196,7 @@
 
 - (UIImage *) _toolbarBackgroundImage
 {
-    const float colorMask[6] = {
+    const CGFloat colorMask[6] = {
         222, 255, 222, 255, 222, 255
     };
